+++ config/acl.php
@@ -10,4 +10,9 @@ return [
      * Permission class used for ACL.
      */
     'permission' => \Yajra\Acl\Models\Permission::class,
+
+    /**
+     * User class used for ACL.
+     */
+    'user' => App\User::class,
 ];